*: move the reconcilation logic into a vault dir
