+++ vault-operator/pkg/vault/deploy_vault.go
@@ -1,4 +1,4 @@
-package stub
+package vault
 
 import (
 	"fmt"
