+++ vault-operator/pkg/vault/doc.go
@@ -1,0 +1,2 @@
+// Package vault contains the reconciliation logic for the vault Custom Resource.
+package vault
