+++ vault-operator/pkg/vault/sync_vault.go
@@ -1,4 +1,4 @@
-package stub
+package vault
 
 import (
 	"fmt"
