+++ vault-operator/pkg/vault/tls.go
@@ -1,4 +1,4 @@
-package stub
+package vault
 
 import (
 	"crypto/rsa"
